Changed Kaleidoscopic enchantment type to "all"
Some things (Such as the openblocks anvil and the mana enchanter) check
for EnumEnchantmentType first instead of CanApply(). Because of this,
some things refuse to put Kaleidoscopic enchatments on swords. This
should fix that, and unless somebody is doing something dumb (Not
checking for CanApply() at all) it shouldn't cause any problems.